requirements.txt
asgiref==3.5.2
attrs==21.4.0
beautifulsoup4==4.11.1
coverage==6.2
dj-database-url==1.0.0
Django==4.1
django-bootstrap4==21.1
django-filter==22.1
django-heroku==0.3.1
flake8==3.9.2
gunicorn==20.1.0
importlib-metadata==2.1.3
iniconfig==1.1.1
mccabe==0.6.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==3.0.7
pytest==6.2.5
pytest-cov==2.12.1
python-dotenv==0.20.0
pytz==2021.3
soupsieve==2.3.2.post1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.1.1
whitenoise==5.3.0
zipp==3.6.0
